+++ src/components/Inputs/File/index.ts
@@ -1,4 +1,0 @@
-import File from "./File";
-
-export default File;
-export * from "./File";