+++ src/components/Inputs/FileInput/index.ts
@@ -1,0 +1,4 @@
+import FileInput from "./FileInput";
+
+export default FileInput;
+export * from "./FileInput";