+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v2.3.11
 - **[fix]** - `File` error with helper text
 - **[fix]** - `File` focusable html element
+- **[refactor]** - rename `File` to `FileInput`