+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v2.3.11
-- **[fix]** - `File` error with helper text
-- **[fix]** - `File` focusable html element
+## v2.3.12
+- **[fix]** - `Timeline` add data test props